Separate CSharp and VisualBasic providers; correctly retrieve template for CSharp.
