+++ src/Microsoft.VisualStudio.ProjectSystem.Managed/ProjectSystem/SpecialFileProviders/VisualBasic/VisualBasicAppConfigSpecialFileProvider.cs
@@ -1,20 +1,18 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
-namespace Microsoft.VisualStudio.ProjectSystem.SpecialFileProviders
+namespace Microsoft.VisualStudio.ProjectSystem.SpecialFileProviders.VisualBasic
 {
     /// <summary>
-    ///     Provides a <see cref="ISpecialFileProvider"/> that handles the default 'App.config' file;
-    ///     which contains .NET Framework directives for assembly binding, compatibility and runtime
-    ///     settings.
+    /// Provides a <see cref="ISpecialFileProvider"/> that handles the default 'App.config' file.
     /// </summary>
     [ExportSpecialFileProvider(SpecialFiles.AppConfig)]
-    [AppliesTo(ProjectCapability.DotNet)]
-    internal class AppConfigSpecialFileProvider : AbstractFindByNameSpecialFileProvider
+    [AppliesTo(ProjectCapability.VisualBasic)]
+    internal class VisualBasicAppConfigSpecialFileProvider : AbstractFindByNameSpecialFileProvider
     {
         private readonly ICreateFileFromTemplateService _templateFileCreationService;
 
         [ImportingConstructor]
-        public AppConfigSpecialFileProvider(IPhysicalProjectTree projectTree, ICreateFileFromTemplateService templateFileCreationService)
+        public VisualBasicAppConfigSpecialFileProvider(IPhysicalProjectTree projectTree, ICreateFileFromTemplateService templateFileCreationService)
             : base("App.config", projectTree)
         {
             _templateFileCreationService = templateFileCreationService;
